+++ ffi_internals/src/consumer/consumer_impl.rs
@@ -2,7 +2,7 @@
 //! Generates a wrapping implementation in the consumer's language.
 //!
 
-use crate::impl_internals::impl_ffi::ImplFFI;
+use crate::items::impl_ffi::ImplFFI;
 
 impl ImplFFI {
     /// Generates an appropriate consumer file name for this impl (by joining the trait and type
@@ -10,7 +10,7 @@ impl ImplFFI {
     ///
     #[must_use]
     pub fn consumer_file_name(&self) -> String {
-        format!("{}_{}.swift", self.trait_name, self.type_name)
+        format!("{}_{}.swift", self.impl_description, self.type_name)
     }
 
     /// Generates an implementation for the consumer's type so that they'll be able to call it like