+++ ffi_internals/src/items/fn_ffi.rs
@@ -4,7 +4,7 @@
 //!
 
 use crate::{
-    parsing::{FieldAttributes, TypeAttributes},
+    parsing::{FieldAttributes, FnAttributes, TypeAttributes},
     type_ffi::{Context, TypeFFI, TypeIdentifier},
 };
 use lazy_static::__Deref;
@@ -57,13 +57,11 @@ pub struct FnFFIInputs<'a> {
     /// The impl's parsed data structure.
     ///
     pub method: &'a ImplItemMethod,
-    /// Any types referenced in the impl that should be passed through the FFI without wrapping,
-    /// such as numerics or `repr(C)` enums/structs.
+
+    /// Data from helper attributes on this function.
     ///
-    pub raw_types: Vec<Ident>,
-    /// The type of `Self` in this impl.
-    ///
-    pub self_type: Ident,
+    pub fn_attributes: &'a FnAttributes,
+
     /// A map of local aliases, where the key is the newtype's identifier and the value is the
     /// underlying type.
     ///
@@ -99,8 +97,7 @@ impl<'a> From<FnFFIInputs<'a>> for FnFFI {
                     syn::FnArg::Typed(arg) => {
                         acc.0.push(FnParameterFFI::from(FnParameterFFIInputs {
                             arg,
-                            raw_types: inputs.raw_types.clone(),
-                            self_type: Some(inputs.self_type.clone()),
+                            fn_attributes: inputs.fn_attributes,
                         }));
                     }
                 }
@@ -114,8 +111,8 @@ impl<'a> From<FnFFIInputs<'a>> for FnFFI {
                 let dealiased = inputs.strip_local_alias(&*ty);
                 Some(TypeFFI::from(TypeAttributes::initial(
                     dealiased,
-                    inputs.raw_types,
-                    Some(inputs.self_type),
+                    inputs.fn_attributes.raw_types.clone(),
+                    Some(inputs.fn_attributes.extend_type.clone()),
                 )))
             }
         };
@@ -129,7 +126,7 @@ impl<'a> From<FnFFIInputs<'a>> for FnFFI {
     }
 }
 
-impl From<(&ItemFn, Vec<Ident>)> for FnFFI {
+impl From<(&ItemFn, &FnAttributes)> for FnFFI {
     /// Converts a tuple of `syn::ItemFn` (the function to build an FFI for) and `Vec<Ident>` (a
     /// collection of raw types that can be exposed directly) to a `FnFFI`.
     ///
@@ -138,8 +135,8 @@ impl From<(&ItemFn, Vec<Ident>)> for FnFFI {
     /// captures additional information available in the impl that may be necessary to build the
     /// FFI function.
     ///
-    fn from(data: (&ItemFn, Vec<Ident>)) -> Self {
-        let (method, raw_types) = data;
+    fn from(data: (&ItemFn, &FnAttributes)) -> Self {
+        let (method, fn_attributes) = data;
         let fn_name = method.sig.ident.clone();
         let (arguments, receiver) = method.sig.inputs.iter().fold(
             (Vec::<FnParameterFFI>::new(), FnReceiver::None),
@@ -155,8 +152,7 @@ impl From<(&ItemFn, Vec<Ident>)> for FnFFI {
                     syn::FnArg::Typed(arg) => {
                         acc.0.push(FnParameterFFI::from(FnParameterFFIInputs {
                             arg,
-                            raw_types: raw_types.clone(),
-                            self_type: None,
+                            fn_attributes,
                         }));
                     }
                 }
@@ -168,8 +164,8 @@ impl From<(&ItemFn, Vec<Ident>)> for FnFFI {
             syn::ReturnType::Default => None,
             syn::ReturnType::Type(_token, ty) => Some(TypeFFI::from(TypeAttributes::initial(
                 *ty.clone(),
-                raw_types,
-                None,
+                fn_attributes.raw_types.clone(),
+                Some(fn_attributes.extend_type.clone()),
             ))),
         };
 
@@ -237,11 +233,6 @@ impl FnFFI {
                 let name = arg.name.clone();
                 let ty = arg.native_type_data.ffi_type(None, Context::Argument);
                 let signature_parameter = quote!(#name: #ty, );
-                // TODO: This assumes a collection type should always be dereferenced to a slice
-                // and borrowed when passed to the native function, which is not necessarily the
-                // case. We should be able to figure that out from the syn collection types...we
-                // just need to support them more completely instead of stripping down to "is a
-                // collection".
                 let symbols = if arg.native_type_data.is_vec {
                     quote!(&*)
                 } else {
@@ -378,12 +369,10 @@ pub struct FnParameterFFIInputs<'a> {
     /// The parameter argument, as in `foo: i32`.
     ///
     arg: &'a PatType,
-    /// Any types that we should expose directly through the FFI.
+
+    /// Data from helper attributes on this function.
     ///
-    raw_types: Vec<Ident>,
-    /// The type of `self` in this context, if available.
-    ///
-    self_type: Option<Ident>,
+    fn_attributes: &'a FnAttributes,
 }
 
 impl<'a> From<FnParameterFFIInputs<'a>> for FnParameterFFI {
@@ -396,10 +385,17 @@ impl<'a> From<FnParameterFFIInputs<'a>> for FnParameterFFI {
                 "Anonymous parameter (not allowed in Rust 2018): {:?}"
             );
         };
+        // If `inputs.arg.ty` is a generic and the appropriate concrete type was provided in the
+        // attributes, use the concrete type as the type of the generated FFI.
+        let concrete_type = inputs
+            .fn_attributes
+            .generics
+            .get_key_value(&*inputs.arg.ty)
+            .map_or(*inputs.arg.ty.clone(), |(_, value)| value.clone());
         let native_type_data = TypeFFI::from(TypeAttributes::initial(
-            *inputs.arg.ty.clone(),
-            inputs.raw_types,
-            inputs.self_type,
+            concrete_type,
+            inputs.fn_attributes.raw_types.clone(),
+            Some(inputs.fn_attributes.extend_type.clone()),
         ));
         Self {
             name,